requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-dateutil>=2.8.2
lxml>=4.9.0

# YAML support
pyyaml>=6.0

# CLI support
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
